Implementation to use pound rather than pence amounts and pass failing tests.

+++ src/main/scala/com/meliore/shopping/CheckoutCalculator.scala
@@ -2,6 +2,6 @@ package com.meliore.shopping
 
 class CheckoutCalculator {
 
-  def calculateTotal(fruits: List[Fruit]): Int =
-    fruits.foldLeft(0)(_ + _.cost)
+  def calculateTotal(fruits: List[Fruit]): Double =
+    fruits.foldLeft(0.0)(_ + _.cost)
 }
